Modularised code further
