+++ data/duke.txt
@@ -1,2 +1,3 @@
 T,1,something else
 E,1,something else,2001-09-09 20:30
+T,1,something
