+++ src/main/java/Storage.java
@@ -1,0 +1,40 @@
+import java.io.File;
+import java.io.FileWriter;
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Scanner;
+
+public class Storage {
+
+    public static void writeToFile(List<Task> tasks) throws IOException {
+        int i = 1;
+        String lines = "";
+        for(Task task: tasks) {
+            lines += task.fileWriteString();
+            if(i != tasks.size()) {
+                lines += "\n";
+            }
+            i++;
+        }
+        FileWriter fileWriter = new FileWriter("data/duke.txt");
+        fileWriter.write(lines);
+        fileWriter.close();
+    }
+
+    public static List<String> readFile() {
+        File file = new File("data/duke.txt");
+        List<String> tasks = new ArrayList();
+        try {
+            Scanner readFile = new Scanner(file);
+            while(readFile.hasNext()) {
+                String task = readFile.nextLine();
+                tasks.add(task);
+            }
+        } catch(Exception err) {
+            System.out.println(err);
+        }
+        return tasks;
+    }
+
+}
